+++ CHANGELOG.md
@@ -7,11 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
 - Added `app.extraContainers` to allow injecting additional sidecar containers into the Deployment.
 - Added `app.envFrom` to allow configuring container `envFrom` sources.
 - Added `app.extraEnv` to allow adding additional container environment entries. [#796](https://github.com/stac-utils/stac-fastapi-elasticsearch-opensearch/pull/796)
-
-### Added
 
 ### Changed
 
@@ -27,6 +27,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Updated
 
 - Updated stac-fastapi parent dependencies from v6.3.0 -> v6.3.2 [#786](https://github.com/stac-utils/stac-fastapi-elasticsearch-opensearch/pull/786)
+- Updated stac-fastapi parent dependencies from v6.3.2 -> v6.4.1 [#812](https://github.com/stac-utils/stac-fastapi-elasticsearch-opensearch/pull/812)
+- Update sort extension to use new conformance classes in app.py for search, collection search, and item search endpoints [#812](https://github.com/stac-utils/stac-fastapi-elasticsearch-opensearch/pull/812)
+
 
 ## [v6.19.0] - 2025-06-23